# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Netraksh API — phone-OTP auth, fraud-number reporting, and trust scoring
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/stats:
    get:
      operationId: adminStats
      tags: [admin]
      summary: Dashboard metrics overview
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        default:
          $ref: "#/components/responses/Error"
  /admin/reports:
    get:
      operationId: adminListReports
      tags: [admin]
      summary: List reports for moderation
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: phone
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminReportListResponse"
        default:
          $ref: "#/components/responses/Error"
  /admin/reports/{id}:
    patch:
      operationId: adminUpdateReport
      tags: [admin]
      summary: Update a report's moderation status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReportStatusRequest"
      responses:
        "200":
          description: Updated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminReport"
        default:
          $ref: "#/components/responses/Error"
  /admin/numbers/{phone}/verify:
    post:
      operationId: adminVerifyNumber
      tags: [admin]
      summary: Mark a number as verified scam (or clear it)
      security:
        - bearerAuth: []
      parameters:
        - name: phone
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyNumberRequest"
      responses:
        "200":
          description: Updated reputation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NumberReputation"
        default:
          $ref: "#/components/responses/Error"
...
  /admin/fraud-map:
    get:
      operationId: adminFraudMap
      tags: [admin]
      summary: Fraud report volume aggregated by Indian state
      description: >-
        Scam/fraud report volume grouped by Indian state from live community
        reports only, ranked highest first.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Fraud volume by state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudMapResponse"
        default:
          $ref: "#/components/responses/Error"
...
  /admin/broadcasts:
    get:
      operationId: adminListBroadcasts
      tags: [admin]
      summary: List past push broadcasts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Broadcast history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BroadcastList"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: adminSendBroadcast
      tags: [admin]
      summary: Send a push broadcast to all registered devices
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendBroadcastRequest"
      responses:
        "200":
          description: The broadcast that was sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Broadcast"
        default:
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    AdminReport:
      type: object
      properties:
        id:
          type: string
        phone:
          type: string
        categoryKey:
          type: string
        description:
          type: string
        incidentDate:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
        reporterId:
          type: string
        reporterName:
          type: string
          nullable: true
        reporterPhone:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        verifiedScam:
          type: boolean
      required:
        - id
        - phone
        - categoryKey
        - description
        - status
        - reporterId
        - createdAt
        - updatedAt
        - verifiedScam
    AdminReportListResponse:
      type: object
      properties:
        reports:
          type: array
          items:
            $ref: "#/components/schemas/AdminReport"
        total:
          type: integer
      required:
        - reports
        - total
    AdminStats:
      type: object
      properties:
        totalUsers:
          type: integer
        premiumUsers:
          type: integer
        fraudReports:
          type: integer
        blockedNumbers:
          type: integer
        revenuePaise:
          type: integer
        dailyActiveUsers:
          type: integer
      required:
        - totalUsers
        - premiumUsers
        - fraudReports
        - blockedNumbers
        - revenuePaise
        - dailyActiveUsers
...
    FraudMapState:
      type: object
      properties:
        state:
          type: string
        code:
          type: string
          description: Short state code (e.g. UP, MH, KA).
        reports:
          type: integer
      required:
        - state
        - code
        - reports
    FraudMapResponse:
      type: object
      properties:
        states:
          type: array
          items:
            $ref: "#/components/schemas/FraudMapState"
        total:
          type: integer
      required:
        - states
        - total
...
    UpdateReportStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, verified, rejected, spam]
      required:
        - status
    VerifyNumberRequest:
      type: object
      properties:
        verifiedScam:
          type: boolean
      required:
        - verifiedScam
    NumberReputation:
      type: object
      properties:
        phone:
          type: string
        reportCount:
          type: integer
        verifiedScam:
          type: boolean
        lastReportedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - phone
        - reportCount
        - verifiedScam
...
    Broadcast:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        body:
          type: string
        recipientCount:
          type: integer
        successCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - body
        - recipientCount
        - successCount
        - createdAt
    BroadcastList:
      type: object
      properties:
        broadcasts:
          type: array
          items:
            $ref: "#/components/schemas/Broadcast"
      required:
        - broadcasts
    SendBroadcastRequest:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
      required:
        - title
        - body