# lib/api-zod/src/generated/types/broadcast.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Netraksh API — phone-OTP auth, fraud-number reporting, and trust scoring
 * OpenAPI spec version: 0.1.0
 */

export interface Broadcast {
  id: string;
  title: string;
  body: string;
  recipientCount: number;
  successCount: number;
  createdAt: Date;
}