# lib/api-zod/src/generated/types/broadcastList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Netraksh API — phone-OTP auth, fraud-number reporting, and trust scoring
 * OpenAPI spec version: 0.1.0
 */
import type { Broadcast } from './broadcast';

export interface BroadcastList {
  broadcasts: Broadcast[];
}